Rename API and KPI modules to snake_case
